Moved simple_packet into separate file to break include cycle. Closes #835
